requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
mysqlclient==2.2.7
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
